games-emulation/yuzu: provide fallback if gamedb is down

+++ games-emulation/yuzu/yuzu-0.1491.ebuild
@@ -68,7 +68,8 @@ PATCHES=(
 )
 
 pkg_setup() {
-	wget -O "${T}/compatibility_list.json" https://api.yuzu-emu.org/gamedb/ || die
+	wget -t 1 --timeout=15 -O "${T}/compatibility_list.json" https://api.yuzu-emu.org/gamedb/ ||
+		rm -f "${T}/compatibility_list.json"
 }
 
 src_prepare() {
@@ -88,6 +89,10 @@ src_prepare() {
 	sed -re 's/set\(CAN_BUILD_NX_TZDB.*/set(CAN_BUILD_NX_TZDB false)/' -i externals/nx_tzdb/CMakeLists.txt || die
 	cmake_src_prepare
 	mkdir -p "${BUILD_DIR}/dist/compatibility_list" || die
+	if ! [ -f "${T}/compatibility_list.json" ]; then
+		einfo 'Using fallback compatibility list'
+		gunzip < "${FILESDIR}/${PN}-fallback-compat.json.gz" > "${T}/compatibility_list.json" || die
+	fi
 	mv -f "${T}/compatibility_list.json" "${BUILD_DIR}/dist/compatibility_list/compatibility_list.json" || die
 }
